fix: remove prompt stuff
