+++ lilypad/__init__.py
@@ -8,7 +8,6 @@ from . import server as server
 from ._configure import configure
 from .generations import generation
 from .messages import Message
-from .prompts import prompt
 from .response_models import response_model
 from .tools import tool
 
@@ -24,7 +23,6 @@ __all__ = [
     "evals",
     "generation",
     "Message",
-    "prompt",
     "server",
     "tool",
     "response_model",
