+++ lilypad/_utils/__init__.py
@@ -3,14 +3,13 @@
 from .call_safely import call_safely
 from .closure import Closure, DependencyInfo, get_qualified_name
 from .config import load_config
-from .functions import create_mirascope_call, inspect_arguments
+from .functions import inspect_arguments
 from .middleware import create_mirascope_middleware, encode_gemini_part
 
 __all__ = [
     "Closure",
     "call_safely",
     "DependencyInfo",
-    "create_mirascope_call",
     "create_mirascope_middleware",
     "encode_gemini_part",
     "get_qualified_name",
