+++ lilypad/_utils/functions.py
@@ -3,27 +3,19 @@
 import inspect
 import os
 import types
-from collections.abc import AsyncIterable, Callable, Coroutine, Iterable
-from functools import partial, wraps
+from collections.abc import Callable
+from functools import partial
 from importlib import import_module
 from typing import (
     Any,
     ParamSpec,
     TypeVar,
     Union,
-    cast,
     get_args,
     get_origin,
-    get_type_hints,
-    overload,
 )
 
-from mirascope.core import base as mb
-from pydantic import BaseModel
-
-from ..messages import Message
-from ..server.models import Provider
-from ..server.schemas import PromptCreate, PromptPublic
+from ..server.models.spans import Provider
 
 _P = ParamSpec("_P")
 _R = TypeVar("_R")
@@ -135,166 +127,4 @@ def _construct_call_decorator(
     )
 
 
-@overload
-def create_mirascope_call(
-    fn: Callable[_P, Coroutine[Any, Any, _R]],
-    prompt: PromptPublic | PromptCreate,
-    provider: Provider,
-    model: str,
-    trace_decorator: Callable | None,
-) -> Callable[_P, Coroutine[Any, Any, _R]]: ...
-
-
-@overload
-def create_mirascope_call(
-    fn: Callable[_P, _R],
-    prompt: PromptPublic | PromptCreate,
-    provider: Provider,
-    model: str,
-    trace_decorator: Callable | None,
-) -> Callable[_P, _R]: ...
-
-
-def create_mirascope_call(
-    fn: Callable[_P, _R] | Callable[_P, Coroutine[Any, Any, _R]],
-    prompt: PromptPublic | PromptCreate,
-    provider: Provider,
-    model: str,
-    trace_decorator: Callable | None,
-) -> Callable[_P, _R] | Callable[_P, Coroutine[Any, Any, _R]]:
-    """Returns the constructed Mirascope call function."""
-    if not trace_decorator:
-        trace_decorator = lambda x: x  # noqa: E731
-
-    call_decorator = _construct_call_decorator(fn, provider, model)
-    return_type = get_type_hints(fn).get("return", type(None))
-    if inspect.iscoroutinefunction(fn):
-
-        @mb.prompt_template(prompt.template)
-        @wraps(fn)
-        async def prompt_template_async(
-            *args: _P.args, **kwargs: _P.kwargs
-        ) -> mb.BaseDynamicConfig:
-            return {"call_params": prompt.call_params}
-
-        @wraps(fn)
-        async def inner_async(*args: _P.args, **kwargs: _P.kwargs) -> _R:
-            if return_type is str:
-                traced_call = trace_decorator(call_decorator()(prompt_template_async))
-                return (await traced_call(*args, **kwargs)).content
-            origin_return_type = get_origin(return_type)
-            if origin_return_type is AsyncIterable and get_args(return_type) == (str,):
-                traced_call = trace_decorator(
-                    call_decorator(stream=True)(prompt_template_async)
-                )
-
-                async def iterable() -> AsyncIterable[str]:
-                    async for chunk, _ in await traced_call(*args, **kwargs):
-                        yield chunk.content
-
-                return cast(_R, iterable())
-            elif origin_return_type is Message:
-                traced_call = trace_decorator(
-                    call_decorator(tools=list(get_args(return_type)))(
-                        prompt_template_async
-                    )
-                )
-                return cast(_R, Message(await traced_call(*args, **kwargs)))  # pyright: ignore [reportAbstractUsage] # pyright: ignore [reportAbstractUsage]
-            elif (
-                origin_return_type is AsyncIterable
-                and len(iter_args := get_args(return_type)) == 1
-                and issubclass((response_model := iter_args[0]), BaseModel)
-            ):
-                traced_call = trace_decorator(
-                    call_decorator(response_model=response_model, stream=True)(
-                        prompt_template_async
-                    )
-                )
-                return cast(_R, await traced_call(*args, **kwargs))
-            elif (
-                inspect.isclass(origin_return_type)
-                and issubclass(origin_return_type, Message)
-                or (
-                    inspect.isclass(return_type)
-                    and type(return_type) is not types.GenericAlias
-                    and issubclass(return_type, Message)
-                )
-            ):
-                traced_call = trace_decorator(call_decorator()(prompt_template_async))
-                return cast(_R, Message(await traced_call(*args, **kwargs)))  # pyright: ignore [reportAbstractUsage]
-            elif mb._utils.is_base_type(return_type) or (
-                inspect.isclass(return_type)
-                and type(return_type) is not types.GenericAlias
-                and issubclass(return_type, BaseModel)
-            ):
-                traced_call = trace_decorator(
-                    call_decorator(response_model=return_type)(prompt_template_async)
-                )
-                return cast(_R, await traced_call(*args, **kwargs))
-            else:
-                raise ValueError(f"Unsupported return type `{return_type}`.")
-
-        return inner_async
-    else:
-
-        @mb.prompt_template(prompt.template)
-        @wraps(fn)
-        def prompt_template(
-            *args: _P.args, **kwargs: _P.kwargs
-        ) -> mb.BaseDynamicConfig:
-            return {"call_params": prompt.call_params}
-
-        @wraps(fn)
-        def inner(*args: _P.args, **kwargs: _P.kwargs) -> _R:
-            if return_type is str:
-                traced_call = trace_decorator(call_decorator()(prompt_template))
-                return traced_call(*args, **kwargs).content
-            origin_return_type = get_origin(return_type)
-            if origin_return_type is Iterable and get_args(return_type) == (str,):
-                traced_call = trace_decorator(
-                    call_decorator(stream=True)(prompt_template)
-                )
-
-                def iterable() -> Iterable[str]:
-                    for chunk, _ in traced_call(*args, **kwargs):
-                        yield chunk.content
-
-                return cast(_R, iterable())
-            elif origin_return_type is Message:
-                traced_call = trace_decorator(
-                    call_decorator(tools=list(get_args(return_type)))(prompt_template)
-                )
-                return cast(_R, Message(traced_call(*args, **kwargs)))  # pyright: ignore [reportAbstractUsage]
-            elif (
-                origin_return_type is Iterable
-                and len(iter_args := get_args(return_type)) == 1
-                and issubclass((response_model := iter_args[0]), BaseModel)
-            ):
-                traced_call = trace_decorator(
-                    call_decorator(response_model=response_model, stream=True)(
-                        prompt_template
-                    )
-                )
-                return cast(_R, traced_call(*args, **kwargs))
-            elif (
-                inspect.isclass(origin_return_type)
-                and issubclass(origin_return_type, Message)
-                or inspect.isclass(return_type)
-                and issubclass(return_type, Message)
-            ):
-                traced_call = trace_decorator(call_decorator()(prompt_template))
-                return cast(_R, Message(traced_call(*args, **kwargs)))  # pyright: ignore [reportAbstractUsage]
-            elif mb._utils.is_base_type(return_type) or issubclass(
-                return_type, BaseModel
-            ):
-                traced_call = trace_decorator(
-                    call_decorator(response_model=return_type)(prompt_template)
-                )
-                return cast(_R, traced_call(*args, **kwargs))
-            else:
-                raise ValueError(f"Unsupported return type `{return_type}`.")
-
-        return inner
-
-
-__all__ = ["inspect_arguments", "create_mirascope_call"]
+__all__ = ["inspect_arguments"]
