+++ lilypad/server/models/projects.py
@@ -12,7 +12,6 @@ if TYPE_CHECKING:
     from .api_keys import APIKeyTable
     from .generations import GenerationTable
     from .organizations import OrganizationTable
-    from .prompts import PromptTable
     from .response_models import ResponseModelTable
     from .tools import ToolTable
 
@@ -31,9 +30,6 @@ class ProjectTable(_ProjectBase, BaseOrganizationSQLModel, table=True):
         UniqueConstraint("organization_uuid", "name", name="unique_project_name"),
     )
     generations: list["GenerationTable"] = Relationship(
-        back_populates="project", cascade_delete=True
-    )
-    prompts: list["PromptTable"] = Relationship(
         back_populates="project", cascade_delete=True
     )
     response_models: list["ResponseModelTable"] = Relationship(
