+++ lilypad/server/models/table_names.py
@@ -3,7 +3,6 @@
 PROJECT_TABLE_NAME = "projects"
 SPAN_TABLE_NAME = "spans"
 GENERATION_TABLE_NAME = "generations"
-PROMPT_TABLE_NAME = "prompts"
 DEVICE_CODE_TABLE_NAME = "device_codes"
 USER_TABLE_NAME = "users"
 ORGANIZATION_TABLE_NAME = "organizations"
