+++ lilypad/server/schemas/generations.py
@@ -3,7 +3,6 @@
 from uuid import UUID
 
 from ..models.generations import _GenerationBase
-from .prompts import PromptPublic
 from .response_models import ResponseModelPublic
 from .tool import ToolPublic
 
@@ -16,6 +15,5 @@ class GenerationPublic(_GenerationBase):
     """Generation public model."""
 
     uuid: UUID
-    prompt: PromptPublic | None = None
     response_model: ResponseModelPublic | None = None
     tools: list[ToolPublic] | None = None
