+++ lilypad/server/schemas/projects.py
@@ -5,7 +5,6 @@ from uuid import UUID
 
 from ..models.projects import _ProjectBase
 from .generations import GenerationPublic
-from .prompts import PromptPublic
 from .response_models import ResponseModelPublic
 
 
@@ -20,6 +19,5 @@ class ProjectPublic(_ProjectBase):
 
     uuid: UUID
     generations: list[GenerationPublic] = []
-    prompts: list[PromptPublic] = []
     response_models: list[ResponseModelPublic] = []
     created_at: datetime
