+++ lilypad/server/services/__init__.py
@@ -6,7 +6,6 @@ from .generations import GenerationService
 from .organization_invites import OrganizationInviteService
 from .organizations import OrganizationService
 from .projects import ProjectService
-from .prompts import PromptService
 from .spans import SpanService
 from .user_organizations import UserOrganizationService
 from .users import UserService
