+++ tests/test_generations.py
@@ -366,7 +366,7 @@ def test_sync_managed_generation(
     """
     with (
         patch(
-            "lilypad.generations.LilypadClient.get_generation_by_args_types",
+            "lilypad.generations.LilypadClient.get_generation_by_signature",
             return_value=dummy_generation_instance,
         ),
         patch("lilypad.generations.llm.call") as mock_llm_call,
@@ -397,7 +397,7 @@ async def test_async_managed_generation(
     """
     with (
         patch(
-            "lilypad.generations.LilypadClient.get_generation_by_args_types",
+            "lilypad.generations.LilypadClient.get_generation_by_signature",
             return_value=dummy_generation_instance,
         ),
         patch("lilypad.generations.llm.call") as mock_llm_call,
